# `ArgIteratorWindows`: Cleanup and some optimizations